+++ .env
@@ -1,9 +1,10 @@
 SERVER_ADDRESS=localhost:8080
 BASE_URL=localhost:8080
 FILE_STORAGE_PATH=short-url-db.json
-DATABASE_DSN="host=localhost user=shortener password=shortener dbname=shortener sslmode=disable"
+#DATABASE_DSN="host=localhost user=shortener password=shortener dbname=shortener sslmode=disable"
 JWT_SECRET=11sb6GFNBhas86BjjndHJBROcxd4adrmc
 PPROF_HOST=localhost:8888
 ENABLE_HTTPS=true
 #CONFIG=config.json
-#TRUSTED_SUBNET=127.0.0.1
+TRUSTED_SUBNET=127.0.0.1
+GRPC_PORT=3200
