+++ internal/app/server/server.go
@@ -20,17 +20,21 @@ import (
 	pb "github.com/MagicNetLab/ya-practicum-shortener/pkg/shortener_grpc"
 )
 
-var runServers []*http.Server
+const (
+	certFileName = "cert.pem"
+	keyFileName  = "key.pem"
+)
 
 // Run запуск сервера
 func Run(configurator configurator) {
 	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGINT, syscall.SIGTERM, syscall.SIGQUIT)
 	defer stop()
 
+	var runServers []*http.Server
 	if configurator.IsEnableHTTPS() {
-		runHTTPSServer(configurator)
+		runServers = runHTTPSServer(configurator)
 	} else {
-		runHTTPServer(configurator)
+		runServers = runHTTPServer(configurator)
 	}
 
 	grpcServ := runGRPCServer(configurator)
@@ -76,7 +80,8 @@ func getListeners() listeners {
 	return l
 }
 
-func runHTTPServer(configurator configurator) {
+func runHTTPServer(configurator configurator) []*http.Server {
+	var runServers []*http.Server
 	listen := getListeners()
 
 	for h, l := range listen {
@@ -121,13 +126,12 @@ func runHTTPServer(configurator configurator) {
 		args := map[string]interface{}{"url": "http://" + pprofHost}
 		logger.Info("pprof starting", args)
 	}
+
+	return runServers
 }
 
-func runHTTPSServer(configurator configurator) {
-	const (
-		certFileName = "cert.pem"
-		keyFileName  = "key.pem"
-	)
+func runHTTPSServer(configurator configurator) []*http.Server {
+	var runServers []*http.Server
 
 	if !hasTLSCertsExists(certFileName, keyFileName) {
 		err := createTLSCerts(certFileName, keyFileName)
@@ -181,11 +185,13 @@ func runHTTPSServer(configurator configurator) {
 		args := map[string]interface{}{"url": "https://" + pprofHost}
 		logger.Info("pprof starting", args)
 	}
+
+	return runServers
 }
 
 func runGRPCServer(configurator configurator) *grpc.Server {
 	// определяем порт для сервера
-	listen, err := net.Listen("tcp", ":3200")
+	listen, err := net.Listen("tcp", ":"+configurator.GetGRPCPort())
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -212,4 +218,35 @@ func runGRPCServer(configurator configurator) *grpc.Server {
 	}()
 
 	return s
+}
+
+func runPProfServer(configurator configurator) *http.Server {
+	pprofHost := configurator.GetPProfHost()
+	pprofServer := &http.Server{
+		Addr:    pprofHost,
+		Handler: nil,
+	}
+
+	if configurator.IsEnableHTTPS() {
+		go func(s *http.Server) {
+			err := s.ListenAndServeTLS(certFileName, keyFileName)
+			if err != nil && err != http.ErrServerClosed {
+				args := map[string]interface{}{"error": err}
+				logger.Fatal("failed starting pprof server", args)
+			}
+		}(pprofServer)
+	} else {
+		go func(serv *http.Server) {
+			err := serv.ListenAndServe()
+			if err != nil && err != http.ErrServerClosed {
+				args := map[string]interface{}{"error": err.Error()}
+				logger.Error("Failed start pprof server", args)
+			}
+		}(pprofServer)
+	}
+
+	args := map[string]interface{}{"url": "http://" + pprofHost}
+	logger.Info("pprof starting", args)
+
+	return pprofServer
 }
