+++ internal/service/user/user_test.go
@@ -1,0 +1,16 @@
+package user
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func Test_randString(t *testing.T) {
+	l := 5
+
+	str, err := randString(l)
+	assert.NoError(t, err)
+
+	assert.Len(t, str, l)
+}
